requirements.txt
asgiref==3.4.1
async-generator==1.10
attrs==21.2.0
autopep8==1.6.0
certifi==2021.10.8
cffi==1.15.0
colorama==0.4.4
coverage==6.2
cryptography==36.0.0
Django==4.0
django-debug-toolbar==3.2.4
django-jazzmin==2.6.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2
docopt==0.6.2
drf-spectacular==0.26.2
Faker==10.0.0
flake8==4.0.1
gunicorn==20.1.0
h11==0.12.0
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
jsonschema==4.17.3
mccabe==0.6.1
mypy==0.910
mypy-extensions==0.4.3
outcome==1.1.0
packaging==21.3
parameterized==0.8.1
Pillow==8.4.0
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
Pygments==2.15.1
PyJWT==2.7.0
pyOpenSSL==21.0.0
pyparsing==3.0.6
pyrsistent==0.19.3
pytest==6.2.5
pytest-django==4.5.2
pytest-dotenv==0.5.2
pytest-watch==4.2.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
trio==0.19.0
trio-websocket==0.9.2
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
watchdog==2.1.6
whitenoise==5.3.0
wsproto==1.0.0
